Update README.md and created Ejercicios/README.md
The content of README.md has been divided into 2 files with the same
name: the first one refers to the README of the repository, and contains
generic information about it; the second one, in Ejercicios/, acts as an
introduction to Vim (old file Introducción deleted in <commit>).

"README.md" was used for the web to display its contents.